requirements.txt
spotipy>=2.23.0
python-dotenv>=1.0.0
keyboard>=0.13.5
textual>=0.47.1